% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GO.R
\name{GOPlot_barPlot}
\alias{GOPlot_barPlot}
\title{GOPlot_barPlot}
\usage{
GOPlot_barPlot(
  result_GO,
  Clu_levels = NULL,
  title_name = NULL,
  addLine = FALSE,
  space_width = 0.5,
  wrap_length = 40
)
}
\arguments{
\item{wrap_length}{}
}
\value{

}
\description{
GOPlot_barPlot
}
