% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GO.R
\name{GOPlot_barPlot_flip}
\alias{GOPlot_barPlot_flip}
\title{GOPlot_barPlot_flip}
\usage{
GOPlot_barPlot_flip(
  result_GO,
  Clu_levels = NULL,
  title_name = NULL,
  labelInBar = FALSE,
  addLine = FALSE,
  space_width = 0.5,
  wrap_length = 40
)
}
\arguments{
\item{wrap_length}{}
}
\value{

}
\description{
GOPlot_barPlot_flip
}
\examples{

GO_file <- system.file("extdata", "GO.csv", package = "BioinfoPlotTK")
GO_info <- read.csv(GO_file)
p <- GOPlot_barPlot_flip(result_GO = GO_info)


}
