% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_min_max}
\alias{scale_min_max}
\title{scale_min_max}
\usage{
scale_min_max(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
scale_min_max
}
\examples{
mt <- t(matrix(c(rep(0, 5), rnorm(5), rep(2,5)), nrow = 5))
scale_min_max(mt)

}
