% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{volcanoPlot_gene}
\alias{volcanoPlot_gene}
\title{volcanoPlot_gene}
\usage{
volcanoPlot_gene(
  result_diffGene,
  Fold_cutoff = 1,
  padj_cutoff = 0.05,
  gene_select,
  color_palette = c("#669cc7", "#a6a6a6", "#d73c31"),
  ylim_max,
  max.overlaps = 200
)
}
\arguments{
\item{max.overlaps}{}
}
\value{

}
\description{
volcanoPlot_gene
}
